Remove second surrogate indirection
